backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
sqlalchemy
openai
PyJWT
email-validator
